Dodan FamNet.

+++ python_api/__init__.py
@@ -13,7 +13,7 @@ import requests
 import cv2
 import numpy as np
 from PIL import Image
-from workers import train_model, print_task, predict, predict_faster_rcnn
+from workers import train_model, print_task, predict, predict_faster_rcnn, predict_famnet
 import os
 import shutil
 
@@ -174,33 +174,51 @@ class predictHandler(tornado.web.RequestHandler):
     def post(self):
         data = json_decode(self.request.body)
         method = data['method']
-        detection_threshold = float(data['threshold'])
-        label = data['label']
-        images = data['images']
+        if method == 'famnet':
+            self.famnet_predict()
+        else:
+            detection_threshold = float(data['threshold'])
+            label = data['label']
+            images = data['images']
 
-        global pred_images
-        pred_images = []
+            global pred_images
+            pred_images = []
 
-        for img in images:
-            pred_images.append({'image': getPredImage(img), 'name': img})
+            for img in images:
+                pred_images.append({'image': getPredImage(img), 'name': img})
 
-        folder = './predictions'
-        for filename in os.listdir(folder):
-            filepath = os.path.join(folder, filename)
-            try:
-                if os.path.isfile(filepath) or os.path.islink(filepath):
-                    os.unlink(filepath)
-            except Exception as e:
-                print('Failed to delete %s. Reason: %s' % (filepath, e))
+            folder = './predictions'
+            for filename in os.listdir(folder):
+                filepath = os.path.join(folder, filename)
+                try:
+                    if os.path.isfile(filepath) or os.path.islink(filepath):
+                        os.unlink(filepath)
+                except Exception as e:
+                    print('Failed to delete %s. Reason: %s' % (filepath, e))
 
-        if method == 'custom':
-            model = data['model']
-            getModel(model)
-            job = q.enqueue(predict, job_timeout='24h', args=('dmodel.pt', pred_images, detection_threshold, label))
-            self.write({'message': 'success', 'job_id': job.id})
-        elif method == 'faster_rcnn':
-            job = q.enqueue(predict_faster_rcnn, job_timeout='24h', args=(pred_images, detection_threshold, label))
-            self.write({'message': 'success', 'job_id': job.id})
+            if method == 'custom':
+                model = data['model']
+                getModel(model)
+                job = q.enqueue(predict, job_timeout='24h', args=('dmodel.pt', pred_images, detection_threshold, label))
+                self.write({'message': 'success', 'job_id': job.id})
+            elif method == 'faster_rcnn':
+                job = q.enqueue(predict_faster_rcnn, job_timeout='24h', args=(pred_images, detection_threshold, label))
+                self.write({'message': 'success', 'job_id': job.id})
+
+
+    def famnet_predict(self):
+        data = json_decode(self.request.body)
+        annotations = []
+        images = []
+        annotations = data['data']
+
+        for ann in annotations:
+            image_name = ann['image_name']
+            images.append(getImage(image_name))
+
+        print(f'Method: FamNet, Ann. lenght: {len(annotations)}, Imgs. lenght: {len(images)}')
+        job = q.enqueue(predict_famnet, job_timeout='24h', args=(images, annotations,))
+        self.write({'message': 'success', 'job_id': job.id})
 
     def options(self, *args):
         # no body
